pyproject.toml
[project]
name = "marimo-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "marimo[recommended]>=0.15.5",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "pillow>=11.3.0",
    "plotly>=6.3.0",
]
